# server/docs/swagger.yaml
basePath: /
definitions:
...
  community.ForumTopic:
    properties:
      attentionNum:
        description: 关注数
        type: integer
      circleId:
        description: 圈子_编号
        type: integer
      coverImage:
        description: 封面
        type: string
      createdAt:
        description: 创建时间
        type: string
      discussNum:
        description: 讨论数
        type: integer
      hot:
        description: 是否热门：0 否 1 是
        type: integer
      id:
        description: 主键ID
        type: integer
      name:
        description: 名称
        type: string
      reviewStatus:
        description: 审核状态：0 未处理 1 通过，2驳回
        type: integer
      topicGroupId:
        description: 分组id
        type: integer
      type:
        description: 话题类型：0 全局，1圈子
        type: integer
    type: object
  community.ForumTopicGroup:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        description: 名称
        type: string
      sort:
        description: 排序
        type: integer
    type: object
...
  general.UserBrowsingHistory:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      postsId:
        description: 帖子编号
        type: integer
      time:
        description: 浏览时间
        type: string
      userId:
        description: 用户编号
        type: integer
    type: object
...
  request.CreateForumTopicReq:
    properties:
      circleId:
        description: 圈子_编号
        type: integer
      coverImage:
        description: 封面
        type: string
      name:
        description: 名称
        type: string
      topicGroupId:
        description: 分组id
        type: integer
      type:
        description: 话题类型：0 全局，1圈子
        type: integer
    type: object
...
  request.IdDelete:
    properties:
      id:
        description: 编号
        type: integer
    type: object
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  response.DataResult:
    type: object
...
  response.PageResult:
    properties:
      list:
        type: object
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /app/topic/createForumTopic:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建ForumTopic
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateForumTopicReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建ForumTopic
      tags:
      - App_Topic
  /app/topic/deleteForumTopic:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除ForumTopic
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdDelete'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除ForumTopic
      tags:
      - App_Topic
  /app/topic/deleteForumTopicByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除ForumTopic
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除ForumTopic
      tags:
      - App_Topic
  /app/topic/findForumTopic:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编号
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumTopic
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.ForumTopic'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询ForumTopic
      tags:
      - App_Topic
  /app/topic/findForumTopicGroup:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编号
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumTopicGroup
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/community.ForumTopicGroup'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询ForumTopicGroup
      tags:
      - App_Topic
  /app/topic/getForumTopicGroupList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumTopicGroup
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.ForumTopicGroup'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取ForumTopicGroup列表
      tags:
      - App_Topic
  /app/topic/getForumTopicGroupListAll:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumTopicGroup
          schema:
            allOf:
            - $ref: '#/definitions/response.DataResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/community.ForumTopicGroup'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取ForumTopicGroup列表
      tags:
      - App_Topic
  /app/topic/getForumTopicList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 圈子_编号
        in: query
        name: circleId
        type: integer
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 是否热门：0 否 1 是
        in: query
        name: hot
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 审核状态：0 未处理 1 通过，2驳回
        in: query
        name: reviewStatus
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      - description: 分组id
        in: query
        name: topicGroupId
        type: integer
      - description: 话题类型：0 全局，1圈子
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumTopic
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.ForumTopic'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取ForumTopic列表
      tags:
      - App_Topic
  /app/topic/updateForumTopic:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新ForumTopic
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/community.ForumTopic'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新ForumTopic
      tags:
      - App_Topic
...
  /app/userBrowsingHistory/deleteUserBrowsingHistory:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除UserBrowsingHistory
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/general.UserBrowsingHistory'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除UserBrowsingHistory
      tags:
      - APP_UserBrowsingHistory
  /app/userBrowsingHistory/deleteUserBrowsingHistoryByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除UserBrowsingHistory
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除UserBrowsingHistory
      tags:
      - APP_UserBrowsingHistory
  /app/userBrowsingHistory/getUserBrowsingHistoryList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endUpdatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startUpdatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回general.UserBrowsingHistory
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/general.UserBrowsingHistory'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取UserBrowsingHistory列表
      tags:
      - APP_UserBrowsingHistory
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"